\name{display.extract}
\alias{display.extract}
\title{Display method for Dyad Ratios Algorithm Result
}
\description{Displays the latent series output by time-unit.}
\usage{
\method{display}{extract}(out, filename = NULL)
}
\arguments{
  \item{out}{Object of class \code{extract}.}
  \item{filename}{Name of file to which output will be written.}
}
\seealso{
\code{\link{extract}}
}
