% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.extract}
\alias{print.extract}
\title{Print method for extract objects}
\usage{
\method{print}{extract}(x, ...)
}
\arguments{
\item{x}{A \code{extract} object returned by \code{\link{extract}}.}

\item{...}{Ignored.}
}
\description{
Print method for extract objects
}
