\name{DyadRatios-package}
\alias{DyadRatios-package}
\alias{DyadRatios}
\docType{package}
\title{
\packageTitle{DyadRatios}
}
\description{
\packageDescription{DyadRatios}
}
\author{
\packageAuthor{DyadRatios}

Maintainer: \packageMaintainer{DyadRatios}
}
\references{
Stimson, J. A. (2018) \sQuote{The Dyad Ratios Algorithm for Estimating Latent Public Opinion: Estimation, Testing, and Comparison to Other Approaches}, Bulletin of Sociological Methodology/Bulletin de Méthodologie Sociologique, 137–138(1), pp. 201–218. doi: 10.1177/0759106318761614.
}
