% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mood.R
\name{get_mood}
\alias{get_mood}
\title{Get Mood Estimates from Dyad Ratios Algorithm}
\arguments{
\item{out}{An object of class "extract", typically a result of the `extract()` function.}

\item{...}{Other arguments to be passed down (currently unused).}
}
\value{
A data frame containing the period of aggregation and the corresponding latent variable estimates for each dimension.
}
\description{
Retrieve the latent variable estimates from the dyad ratios algorithm produced by the `extract()` function.
}
