\name{plot.extract}
\alias{plot.extract}
\title{Plot method for Dyad Ratios Algorithm results.}
\description{Plot method for Dyad Ratios Algorithm results.}
\usage{
\method{plot}{extract}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{Object of class \code{extract}.}
  \item{...}{Other arguments passed down to \code{plot}.}
}
\seealso{
\code{\link{extract}}
}
