\name{print.extract}
\alias{print.extract}
\title{Print method for Dyad Ratios Algorithm results.}
\description{Print method for Dyad Ratios Algorithm results.}
\usage{
\method{print}{extract}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{Object of class \code{extract}.}
  \item{...}{Other arguments passed down to \code{print}.}
}
\seealso{
\code{\link{extract}}
}
