% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.extract.R
\name{print.extract}
\alias{print.extract}
\title{Print method for extract function}
\arguments{
\item{x}{An object of class "extract", typically a result of the extract function.}

\item{...}{Other arguments to be pased down}
}
\value{
A printed summary of the output from the extract function and the input object returned invisbly.
}
\description{
Print method for extract function
}
\examples{
data(jennings)
dr_out <- extract(varname = jennings$variable, 
                  date = jennings$date, 
                  index = jennings$value, 
                  ncases = jennings$n, 
                  begindt = min(jennings$date), 
                  enddt = max(jennings$date), 
                  npass=1)
print(dr_out)
}
