\name{summary.extract}
\alias{summary.extract}
\alias{summary}
\title{Summary method for Dyad Ratios Algorithm results}
\description{Summary method for Dyad Ratios Algorithm results}
\usage{
\method{summary}{extract}(object, ...)
}
\arguments{
  \item{object}{Object of class \code{extract}.}
  \item{...}{Not implemented}
}
\seealso{
\code{\link{extract}}
}
