% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.extract.R
\name{summary.extract}
\alias{summary.extract}
\title{Summary Method for extract Objects}
\arguments{
\item{object}{An object of class "extract", typically a result of the extract function.}

\item{...}{Other arguments to be passed down (currently unused).}
}
\value{
Printed output of variable loadings and descriptive information.
}
\description{
Prints a summary from objects estimated with the the extract function.
}
\examples{
data(jennings)
dr_out <- extract(varname = jennings$variable, 
                  date = jennings$date, 
                  index = jennings$value, 
                  ncases = jennings$n, 
                  begindt = min(jennings$date), 
                  enddt = max(jennings$date), 
                  npass=1)
summary(dr_out)
}
